+++ lib/feature/native/domain/models/message_extension.model.dart
@@ -2,35 +2,40 @@
 
 part of './models.devicekit.dart';
 
-extension PayloadX on Payload {
-  bool isInvalid() {
-    return error.code.isEmpty && result.value.isEmpty;
+extension ResultX on Result {
+  bool isEmpty() {
+    return !valuebytes.hasValue();
   }
 
-  bool hasReportedErrors() {
-    return error.code.isNotEmpty;
+  bool isInvalid() {
+    return isEmpty() && !shouldBeEmpty;
   }
 }
 
 extension MessageX on Message {
   //
-  Future<Result> validation() async {
-    final payload = this.payload;
 
-    if (payload.isInvalid()) {
+  bool hasReportedErrors() {
+    return error.code.isNotEmpty;
+  }
+
+  bool isEmptyResult() => result.shouldBeEmpty;
+
+  Future<Result> validation() async {
+    if (result.isInvalid()) {
       throw PlatformException(
         code: 'connection-playload-invalid',
         message: 'Please, check our entrypoint container.',
         details: null,
       );
-    } else if (payload.hasReportedErrors()) {
+    } else if (hasReportedErrors()) {
       throw PlatformException(
-        code: payload.error.code,
-        message: payload.error.message,
-        details: payload.error.details,
+        code: error.code,
+        message: error.message,
+        details: error.details,
       );
     } else {
-      return payload.result;
+      return result;
     }
   }
 }