+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Let there be flutter native mirror.
+
+## 0.0.2
+
+* Ignore registered callback, without throwing an exception.